Move fixtures to conftest, fix style checks
